requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
questionary>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
